Create Wire base class and make optional
Created a base class for the Wire objects, that has most of the
functionality.
There are some virtual functions, which take care of the buss specific
things like, begin/end/setSCL/setSDA.

Also object keeps pointer to the I2C register structure for the buss and
got rid of most statitics.

So far I have left the Wire object independent